BindingPopulation::Entropize() is now a Protecter member of BindingPopulation
BindingPopulation::Entropize() is now called only once all BindingModes are added to Population. This change has been made to prevent multiple unecessary calls to this function when handling outliers. This function recalculates the enthalpy and the entropy of each BindingMode upon each call.

BindingMode and BindingPopulation are now friends of the FastOPTICS class